+++ WebApp/client/src/components/LoginForm/LoginForm.jsx
@@ -22,11 +22,10 @@ const LoginForm = () => {
         Password: data.password,
       })
       .then((response) => {
-        console.log(data);
         Auth.setLoggedUser(data);
       })
       .catch((error) => {
-        addError("Login server error");
+        addError("Login error");
       });
   };
